Add flag to card
